requirements.txt
torch==1.7.1+cpu
torchvision==0.8.2+cpu
Pillow==7.0.0
numpy==1.20.2
